browser-use/pyproject.toml
[project]
name = "browser-user"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.11.7",
    "python-dotenv>=1.0.1",
]
